requirements.txt
# FastAPI and async framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Cohere API client
cohere==5.3.0

# Vector database client
qdrant-client==2.7.0

# Database
psycopg2-binary==2.9.9

# PDF and document parsing
PyPDF2==4.0.1
python-magic==0.4.27

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-multipart==0.0.6
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Logging
structlog==23.2.0

# Optional: For streaming responses
aiofiles==23.2.1
